--- CineHive/src/main/java/com/example/CineHive/repository/GoogleUserRepository.java
@@ -1,12 +1,0 @@
-package com.example.CineHive.repository;
-
-import com.example.CineHive.entity.oauth.GoogleUser;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-public interface GoogleUserRepository extends JpaRepository<GoogleUser, Long> {
-    Optional<GoogleUser> findByMemEmail(String memEmail);
-}
